Extract policy field resolution
